Rename game controller to games (#39)
